refactor api service by market asset position account  todo user spot margin and ins and c2c
