--- src/main/java/com/bybit/api/client/BybitAnnouncementRestClient.java
@@ -1,7 +1,0 @@
-package com.bybit.api.client;
-
-import com.bybit.api.client.domain.announcement.request.AnnouncementInfoRequest;
-
-public interface BybitAnnouncementRestClient {
-    Object getAnnouncementInfo(AnnouncementInfoRequest announcementInfoRequest);
-}
