+++ src/main/java/com/bybit/api/client/BybitApiAccountRestClient.java
@@ -1,0 +1,21 @@
+package com.bybit.api.client;
+
+import com.bybit.api.client.domain.account.AccountDataRequest;
+import com.bybit.api.client.domain.position.PositionDataRequest;
+
+public interface BybitApiAccountRestClient {
+    // Account endpoints
+    Object getWalletBalance(AccountDataRequest walletBalanceRequest);
+    Object upgradeAccountToUTA();
+    Object getAccountBorrowHistory(AccountDataRequest borrowHistoryRequest);
+    Object setAccountCollateralCoin(AccountDataRequest setCollateralCoinRequest);
+    Object getAccountCollateralInfo(AccountDataRequest request);
+    Object getAccountCoinGeeks(AccountDataRequest request);
+    Object getAccountFreeRate(AccountDataRequest getFeeRateRequest);
+    Object getAccountInfo();
+    Object getTransactionLog(AccountDataRequest getTransactionLogRequest);
+    Object setAccountMarginMode(AccountDataRequest request);
+    Object modifyAccountMMP(AccountDataRequest setMMPRequest);
+    Object resetAccountMMP(AccountDataRequest request);
+    Object getAccountMMPState(AccountDataRequest request);
+}
