+++ src/main/java/com/bybit/api/client/BybitApiAsyncMarketDataRestClient.java
@@ -1,0 +1,31 @@
+package com.bybit.api.client;
+
+import com.bybit.api.client.domain.announcement.request.AnnouncementInfoRequest;
+import com.bybit.api.client.domain.market.MarketDataRequest;
+import com.bybit.api.client.domain.trade.BatchOrderRequest;
+import com.bybit.api.client.domain.trade.TradeOrderRequest;
+
+import java.io.IOException;
+import java.util.Map;
+
+public interface BybitApiAsyncMarketDataRestClient {
+    // Market endpoints
+    void getServerTime(BybitApiCallback<Object> callback);
+    void getMarketLinesData(MarketDataRequest marketKlineRequest, BybitApiCallback<Object> callback);
+    void getMarketPriceLinesData(MarketDataRequest marketKlineRequest, BybitApiCallback<Object> callback);
+    void getIndexPriceLinesData(MarketDataRequest marketKlineRequest, BybitApiCallback<Object> callback);
+    void getPremiumIndexPriceLinesData(MarketDataRequest marketKlineRequest, BybitApiCallback<Object> callback);
+    void getInstrumentsInfo(MarketDataRequest instrumentInfoRequest, BybitApiCallback<Object> callback);
+    void getMarketOrderBook(MarketDataRequest marketOrderBookRequest, BybitApiCallback<Object> callback);
+    void getMarketTickers(MarketDataRequest marketDataTickerRequest, BybitApiCallback<Object> callback);
+    void getFundingHistory(MarketDataRequest fundingHistoryRequest, BybitApiCallback<Object> callback);
+    void getRecentTradeData(MarketDataRequest recentTradeRequest, BybitApiCallback<Object> callback);
+    void getOpenInterest(MarketDataRequest openInterestRequest, BybitApiCallback<Object> callback);
+    void getHistoricalVolatility(MarketDataRequest historicalVolatilityRequest, BybitApiCallback<Object> callback);
+    void getInsurance(String coin, BybitApiCallback<Object> callback);
+    void getInsurance(BybitApiCallback<Object> callback);
+    void getRiskLimit(MarketDataRequest marketRiskLimitRequest, BybitApiCallback<Object> callback);
+    void getDeliveryPrice(MarketDataRequest deliveryPriceRequest, BybitApiCallback<Object> callback);
+    void getMarketAccountRatio(MarketDataRequest marketAccountRatioRequest, BybitApiCallback<Object> callback);
+    void getAnnouncementInfo(AnnouncementInfoRequest announcementInfoRequest, BybitApiCallback<Object> callback);
+}
