+++ src/main/java/com/bybit/api/client/BybitApiRestClient.java
@@ -14,38 +14,6 @@ import com.bybit.api.client.domain.spot.marginTrade.*;
 import com.bybit.api.client.domain.user.UserDataRequest;
 
 public interface BybitApiRestClient {
-    // Market Data
-    Object getServerTime();
-    Object getMarketLinesData(MarketDataRequest marketKlineRequest);
-    Object getMarketPriceLinesData(MarketDataRequest marketKlineRequest);
-    Object getIndexPriceLinesData(MarketDataRequest marketKlineRequest);
-    Object getPremiumIndexPriceLinesData(MarketDataRequest marketKlineRequest);
-    Object getInstrumentsInfo(MarketDataRequest instrumentInfoRequest);
-    Object getMarketOrderBook(MarketDataRequest marketOrderBookRequest);
-    Object getMarketTickers(MarketDataRequest marketDataTickerRequest);
-    Object getFundingHistory(MarketDataRequest fundingHistoryRequest);
-    Object getRecentTradeData(MarketDataRequest recentTradeRequest);
-    Object getOpenInterest(MarketDataRequest openInterestRequest);
-    Object getHistoricalVolatility(MarketDataRequest HistoricalVolatilityRequest);
-    Object getInsurance(String coin);
-    Object getInsurance();
-    Object getRiskLimit(MarketDataRequest marketRiskLimitRequest);
-    Object getDeliveryPrice(MarketDataRequest deliveryPriceRequest);
-    Object getMarketAccountRatio(MarketDataRequest marketAccountRatioRequest);
-
-    // Position Data
-    Object getPositionInfo(PositionDataRequest positionListRequest);
-    Object setPositionLeverage(PositionDataRequest setLeverageRequest);
-    Object swithMarginRequest(PositionDataRequest switchMarginRequest);
-    Object switchPositionMode(PositionDataRequest switchPositionModeRequest);
-    Object setTpslMode(PositionDataRequest setTpSlModeRequest);
-    Object setRiskLimit(PositionDataRequest setRiskLimitRequest);
-    Object setTradingStop(PositionDataRequest tradingStopRequest);
-    Object setAutoAddMargin(PositionDataRequest setAutoAddMarginRequest);
-    Object modifyPositionMargin(PositionDataRequest modifyMarginRequest);
-    Object getExecutionList(PositionDataRequest executionHistoryRequest);
-    Object getClosePnlList(PositionDataRequest closePnlHistoryRequest);
-
     // User Data
     Object createSubMember(UserDataRequest subUserRequest);
     Object createSubAPI(UserDataRequest createApiKeyRequest);
@@ -66,47 +34,6 @@ public interface BybitApiRestClient {
     Object getPreUpgradeTransaction(PreUpgradeDataRequest preUpgradeTransactionRequest);
     Object getPreUpgradeOptionDelivery(PreUpgradeDataRequest preUpgradeOptionDeliveryRequest);
     Object getPreUpgradeUsdcSettlement(PreUpgradeDataRequest preUpgradeUsdcSettlementRequest);
-
-    // Account endpoints
-    Object getWalletBalance(AccountDataRequest walletBalanceRequest);
-    Object upgradeAccountToUTA();
-    Object getAccountBorrowHistory(AccountDataRequest borrowHistoryRequest);
-    Object setAccountCollateralCoin(AccountDataRequest setCollateralCoinRequest);
-    Object getAccountCollateralInfo(AccountDataRequest request);
-    Object getAccountCoinGeeks(AccountDataRequest request);
-    Object getAccountFreeRate(AccountDataRequest getFeeRateRequest);
-    Object getAccountInfo();
-    Object getTransactionLog(AccountDataRequest getTransactionLogRequest);
-    Object setAccountMarginMode(AccountDataRequest request);
-    Object modifyAccountMMP(AccountDataRequest setMMPRequest);
-    Object resetAccountMMP(AccountDataRequest request);
-    Object getAccountMMPState(AccountDataRequest request);
-
-    // Asset Endpoints
-    Object getAssetCoinExchangeRecords(AssetDataRequest coinExchangeRecordsRequest);
-    Object getAssetDeliveryRecords(AssetDataRequest deliveryRecordsRequest);
-    Object getAssetUSDCSettlementRecords(AssetDataRequest usdcSettlementRequest);
-    Object getAssetInfo(AssetDataRequest assetInfoRequest);
-    Object getAssetAllCoinsBalance(AssetDataRequest allCoinsBalanceRequest);
-    Object getAssetTransferableCoins(AssetDataRequest request);
-    Object getAssetSingleCoinBalance(AssetDataRequest singleCoinBalanceRequest);
-    Object createAssetInternalTransfer(AssetDataRequest assetInternalTransferRequest);
-    Object getAssetTransferSubUidList();
-    Object createAssetUniversalTransfer(AssetDataRequest assetUniversalTransferRequest);
-    Object getAssetInternalTransferRecords(AssetDataRequest internalTransferRequest);
-    Object getAssetUniversalTransferRecords(AssetDataRequest universalTransferRequest);
-    Object getAssetAllowedDepositCoinInfo(AssetDataRequest allowedDepositCoinRequest);
-    Object setAssetDepositAccount(AssetDataRequest request);
-    Object getAssetDepositRecords(AssetDataRequest assetDepositRecordsRequest);
-    Object getAssetSubMembersDepositRecords(AssetDataRequest assetDepositRecordsRequest);
-    Object getAssetInternalDepositRecords(AssetDataRequest assetDepositRecordsRequest);
-    Object getAssetMasterDepositAddress(AssetDataRequest masterDepositRequest);
-    Object getAssetSubMemberDepositAddress(AssetDataRequest subDepositRequest);
-    Object getAssetCoinInfo(AssetDataRequest request);
-    Object getAssetWithdrawalAmount(AssetDataRequest request);
-    Object getAssetWithdrawalRecords(AssetDataRequest assetWithdrawRecordsRequest);
-    Object cancelAssetWithdraw(AssetDataRequest request);
-    Object createAssetWithdraw(AssetDataRequest assetWithdrawRequest);
 
     // Spot Endpoints
     // Spot Leverage Token
