+++ src/main/java/com/bybit/api/client/BybitMarketDataRestClient.java
@@ -1,0 +1,26 @@
+package com.bybit.api.client;
+
+import com.bybit.api.client.domain.announcement.request.AnnouncementInfoRequest;
+import com.bybit.api.client.domain.market.MarketDataRequest;
+
+public interface BybitMarketDataRestClient {
+    // Market Data
+    Object getServerTime();
+    Object getMarketLinesData(MarketDataRequest marketKlineRequest);
+    Object getMarketPriceLinesData(MarketDataRequest marketKlineRequest);
+    Object getIndexPriceLinesData(MarketDataRequest marketKlineRequest);
+    Object getPremiumIndexPriceLinesData(MarketDataRequest marketKlineRequest);
+    Object getInstrumentsInfo(MarketDataRequest instrumentInfoRequest);
+    Object getMarketOrderBook(MarketDataRequest marketOrderBookRequest);
+    Object getMarketTickers(MarketDataRequest marketDataTickerRequest);
+    Object getFundingHistory(MarketDataRequest fundingHistoryRequest);
+    Object getRecentTradeData(MarketDataRequest recentTradeRequest);
+    Object getOpenInterest(MarketDataRequest openInterestRequest);
+    Object getHistoricalVolatility(MarketDataRequest HistoricalVolatilityRequest);
+    Object getInsurance(String coin);
+    Object getInsurance();
+    Object getRiskLimit(MarketDataRequest marketRiskLimitRequest);
+    Object getDeliveryPrice(MarketDataRequest deliveryPriceRequest);
+    Object getMarketAccountRatio(MarketDataRequest marketAccountRatioRequest);
+    Object getAnnouncementInfo(AnnouncementInfoRequest announcementInfoRequest);
+}
