+++ src/main/java/com/bybit/api/client/impl/BybitMarketDataAsyncRestClientImpl.java
@@ -1,0 +1,200 @@
+package com.bybit.api.client.impl;
+
+import com.bybit.api.client.BybitApiAsyncMarketDataRestClient;
+import com.bybit.api.client.BybitApiAsyncTradeRestClient;
+import com.bybit.api.client.BybitApiCallback;
+import com.bybit.api.client.BybitApiService;
+import com.bybit.api.client.domain.announcement.request.AnnouncementInfoRequest;
+import com.bybit.api.client.domain.market.MarketDataRequest;
+import com.bybit.api.client.domain.trade.BatchOrderRequest;
+import com.bybit.api.client.domain.trade.TradeOrderRequest;
+import com.bybit.api.client.service.BybitJsonConverter;
+
+import java.io.IOException;
+import java.util.Map;
+
+import static com.bybit.api.client.service.BybitApiServiceGenerator.createService;
+
+public class BybitMarketDataAsyncRestClientImpl implements BybitApiAsyncMarketDataRestClient {
+    private final BybitApiService bybitApiService;
+    private final BybitJsonConverter converter = new BybitJsonConverter();
+
+    public BybitMarketDataAsyncRestClientImpl() {
+        bybitApiService = createService(BybitApiService.class);
+    }
+    // Market Data endpoints
+
+    @Override
+    public void getServerTime(BybitApiCallback<Object> callback) {
+        bybitApiService.getServerTime().enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getMarketLinesData(MarketDataRequest marketKlineRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getMarketLinesData(
+                marketKlineRequest.getCategory().getProductTypeId(),
+                marketKlineRequest.getSymbol(),
+                marketKlineRequest.getMarketInterval() == null ? null : marketKlineRequest.getMarketInterval().getIntervalId(),
+                marketKlineRequest.getStart(),
+                marketKlineRequest.getEnd(),
+                marketKlineRequest.getLimit()).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getMarketPriceLinesData(MarketDataRequest marketKlineRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getMarketPriceLinesData(
+                marketKlineRequest.getCategory().getProductTypeId(),
+                marketKlineRequest.getSymbol(),
+                marketKlineRequest.getMarketInterval() == null ? null : marketKlineRequest.getMarketInterval().getIntervalId(),
+                marketKlineRequest.getStart(),
+                marketKlineRequest.getEnd(),
+                marketKlineRequest.getLimit()).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getIndexPriceLinesData(MarketDataRequest marketKlineRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getIndexPriceLinesData(
+                marketKlineRequest.getCategory().getProductTypeId(),
+                marketKlineRequest.getSymbol(),
+                marketKlineRequest.getMarketInterval() == null ? null : marketKlineRequest.getMarketInterval().getIntervalId(),
+                marketKlineRequest.getStart(),
+                marketKlineRequest.getEnd(),
+                marketKlineRequest.getLimit()).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getPremiumIndexPriceLinesData(MarketDataRequest marketKlineRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getPremiumIndexPriceLinesData(
+                marketKlineRequest.getCategory().getProductTypeId(),
+                marketKlineRequest.getSymbol(),
+                marketKlineRequest.getMarketInterval() == null ? null : marketKlineRequest.getMarketInterval().getIntervalId(),
+                marketKlineRequest.getStart(),
+                marketKlineRequest.getEnd(),
+                marketKlineRequest.getLimit()).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getInstrumentsInfo(MarketDataRequest instrumentInfoRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getInstrumentsInfo(
+                instrumentInfoRequest.getCategory().getProductTypeId(),
+                instrumentInfoRequest.getSymbol(),
+                instrumentInfoRequest.getInstrumentStatus() == null ? null : instrumentInfoRequest.getInstrumentStatus().getStatus(),
+                instrumentInfoRequest.getBaseCoin(),
+                instrumentInfoRequest.getLimit(),
+                instrumentInfoRequest.getCursor()
+        ).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getMarketOrderBook(MarketDataRequest marketOrderBookRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getMarketOrderBook(
+                marketOrderBookRequest.getCategory().getProductTypeId(),
+                marketOrderBookRequest.getSymbol(),
+                marketOrderBookRequest.getLimit()
+        ).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getMarketTickers(MarketDataRequest marketDataTickerRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getMarketTickers(
+                marketDataTickerRequest.getCategory().getProductTypeId(),
+                marketDataTickerRequest.getSymbol(),
+                marketDataTickerRequest.getBaseCoin(),
+                marketDataTickerRequest.getExpDate()
+        ).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getFundingHistory(MarketDataRequest fundingHistoryRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getFundingHistory(
+                fundingHistoryRequest.getCategory().getProductTypeId(),
+                fundingHistoryRequest.getSymbol(),
+                fundingHistoryRequest.getStartTime(),
+                fundingHistoryRequest.getEndTime(),
+                fundingHistoryRequest.getLimit()
+        ).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getRecentTradeData(MarketDataRequest recentTradeRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getRecentTradeData(
+                recentTradeRequest.getCategory().getProductTypeId(),
+                recentTradeRequest.getBaseCoin(),
+                recentTradeRequest.getOptionType() == null ? null : recentTradeRequest.getOptionType().getOpType(),
+                recentTradeRequest.getSymbol(),
+                recentTradeRequest.getLimit()
+        ).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getOpenInterest(MarketDataRequest openInterestRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getOpenInterest(
+                openInterestRequest.getCategory().getProductTypeId(),
+                openInterestRequest.getSymbol(),
+                openInterestRequest.getMarketInterval() == null ? null : openInterestRequest.getMarketInterval().getIntervalId(),
+                openInterestRequest.getStartTime(),
+                openInterestRequest.getEndTime(),
+                openInterestRequest.getLimit(),
+                openInterestRequest.getCursor()
+        ).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getHistoricalVolatility(MarketDataRequest historicalVolatilityRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getHistoricalVolatility(
+                historicalVolatilityRequest.getCategory().getProductTypeId(),
+                historicalVolatilityRequest.getBaseCoin(),
+                historicalVolatilityRequest.getOptionPeriod(),
+                historicalVolatilityRequest.getStartTime(),
+                historicalVolatilityRequest.getEndTime()
+        ).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getInsurance(String coin, BybitApiCallback<Object> callback) {
+        bybitApiService.getInsurance(coin).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getInsurance(BybitApiCallback<Object> callback) {
+        bybitApiService.getInsurance().enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getRiskLimit(MarketDataRequest marketRiskLimitRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getRiskLimit(
+                        marketRiskLimitRequest.getCategory().getProductTypeId(),
+                        marketRiskLimitRequest.getSymbol()
+                ).
+                enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getDeliveryPrice(MarketDataRequest deliveryPriceRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getDeliveryPrice(deliveryPriceRequest.getCategory().getProductTypeId(),
+                deliveryPriceRequest.getSymbol(),
+                deliveryPriceRequest.getBaseCoin(),
+                deliveryPriceRequest.getLimit(),
+                deliveryPriceRequest.getCursor()).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getMarketAccountRatio(MarketDataRequest marketAccountRatioRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getMarketAccountRatio(marketAccountRatioRequest.getCategory().getProductTypeId(),
+                marketAccountRatioRequest.getSymbol(),
+                marketAccountRatioRequest.getDataRecordingPeriod() == null ? null : marketAccountRatioRequest.getDataRecordingPeriod().getPeriod(),
+                marketAccountRatioRequest.getLimit()
+        ).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+
+    @Override
+    public void getAnnouncementInfo(AnnouncementInfoRequest announcementInfoRequest, BybitApiCallback<Object> callback) {
+        bybitApiService.getAnouncementInfo(
+                announcementInfoRequest.getLocale() == null ? null : announcementInfoRequest.getLocale().getLanguageSymbol(),
+                announcementInfoRequest.getType() == null ? null : announcementInfoRequest.getType().getAnnouncementType(),
+                announcementInfoRequest.getTag() == null ? null : announcementInfoRequest.getTag().getEnglishTranslation(),
+                announcementInfoRequest.getPage(),
+                announcementInfoRequest.getLimit()
+        ).enqueue(new BybitApiCallbackAdapter<>(callback));
+    }
+}
