+++ src/main/java/com/bybit/api/client/service/BybitApiClientFactory.java
@@ -101,10 +101,10 @@ public class BybitApiClientFactory {
     }
 
     /**
-     * Creates a new synchronous/blocking REST client to Announcement Endpoints
+     * Creates a new synchronous/blocking REST client to Market Data Endpoints
      */
-    public BybitAnnouncementRestClient newAnnouncementRestClient() {
-        return new BybitAnnouncementRestClientImpl();
+    public BybitMarketDataRestClient newMarketDataRestClient() {
+        return new BybitMarketDataRestClientImpl();
     }
 
     /**
@@ -112,6 +112,13 @@ public class BybitApiClientFactory {
      */
     public BybitApiAsyncRestClient newAsyncRestClient() {
         return new BybitApiAsyncRestClientImpl(apiKey, secret);
+    }
+
+    /**
+     * Creates a new asynchronous/non-blocking client to Market Data Endpoints
+     */
+    public BybitApiAsyncMarketDataRestClient newAsyncMarketDataRestClient() {
+        return new BybitMarketDataAsyncRestClientImpl();
     }
 
     /**
@@ -136,5 +143,47 @@ public class BybitApiClientFactory {
 
     public WebsocketClient newWebsocketClient(WebsocketMessageHandler messageHandler) {
         return new WebsocketClientImpl(apiKey, secret, BybitApiConfig.useTestnet ? BybitApiConfig.STREAM_TESTNET_DOMAIN : BybitApiConfig.STREAM_MAINNET_DOMAIN, messageHandler);
+    }
+
+    /**
+     * Creates a new synchronous/blocking REST client to position data
+     */
+    public BybitApiPositionRestClient newPositionRestClient() {
+        return new BybitPositionRestClientImpl(apiKey, secret);
+    }
+
+    /**
+     * Creates a new asynchronous/non-blocking client to position data
+     */
+    public BybitApiAsyncPositionRestClient newAsyncPositionRestClient() {
+        return new BybitApiAsyncPositionRestClientImpl(apiKey, secret);
+    }
+
+    /**
+     * Creates a new synchronous/blocking REST client to Account data
+     */
+    public BybitApiAccountRestClient newAccountRestClient() {
+        return new BybitAccountRestClientImpl(apiKey, secret);
+    }
+
+    /**
+     * Creates a new asynchronous/non-blocking client to Account data
+     */
+    public BybitApiAsyncAccountRestClient newAsyncAccountRestClient() {
+        return new BybitApiAsyncAccountRestClientImpl(apiKey, secret);
+    }
+
+    /**
+     * Creates a new synchronous/blocking REST client to Asset data
+     */
+    public BybitApiAssetRestClient newAssetRestClient() {
+        return new BybitAssetRestClientImpl(apiKey, secret);
+    }
+
+    /**
+     * Creates a new asynchronous/non-blocking client to Asset data
+     */
+    public BybitApiAsyncAssetRestClient newAsyncAssetRestClient() {
+        return new BybitApiAsyncAssetRestClientImpl(apiKey, secret);
     }
 }
