+++ src/test/java/com/bybit/api/examples/http/async/AccountAsyncExample.java
@@ -8,7 +8,7 @@ import com.bybit.api.client.service.BybitApiClientFactory;
 public class AccountAsyncExample {
     public static void main(String[] args) {
         BybitApiClientFactory factory = BybitApiClientFactory.newInstance("YOUR_API_KEY", "YOUR_API_SECRET");
-        var client = factory.newAsyncRestClient();
+        var client = factory.newAsyncAccountRestClient();
 
         // Get wallet balance
         var walletBalanceRequest = AccountDataRequest.builder().accountType(AccountType.UNIFIED).build();
