+++ src/test/java/com/bybit/api/examples/http/async/AssetAsyncExample.java
@@ -10,7 +10,7 @@ import com.bybit.api.client.service.BybitApiClientFactory;
 public class AssetAsyncExample {
     public static void main(String[] args) {
         BybitApiClientFactory factory = BybitApiClientFactory.newInstance("YOUR_API_KEY", "YOUR_API_SECRET");
-        var client = factory.newAsyncRestClient();
+        var client = factory.newAsyncAssetRestClient();
 
         // Get Coin Exchange Records
         var coinExchangeRecordsRequest = AssetDataRequest.builder().build();
