+++ src/test/java/com/bybit/api/examples/http/async/PositionDataAsyncExample.java
@@ -9,7 +9,7 @@ import com.bybit.api.client.service.BybitApiClientFactory;
 public class PositionDataAsyncExample {
     public static void main(String[] args) {
         BybitApiClientFactory factory = BybitApiClientFactory.newInstance("YOUR_API_KEY", "YOUR_API_SECRET");
-        BybitApiAsyncRestClient client = factory.newAsyncRestClient();
+        var client = factory.newAsyncPositionRestClient();
 
         // Get Position Info
         var positionListRequest = PositionDataRequest.builder().category(ProductType.LINEAR).symbol("BTCUSDT").build();
