+++ src/test/java/com/bybit/api/examples/http/sync/AccountExample.java
@@ -11,7 +11,7 @@ import com.bybit.api.client.BybitApiRestClient;
 public class AccountExample {
     public static void main(String[] args) {
         BybitApiClientFactory factory = BybitApiClientFactory.newInstance("YOUR_API_KEY", "YOUR_API_SECRET");
-        BybitApiRestClient client = factory.newRestClient();
+        var client = factory.newAccountRestClient();
 
         // Get wallet balance
         var walletBalanceRequest = AccountDataRequest.builder().accountType(AccountType.UNIFIED).build();
