+++ src/test/java/com/bybit/api/examples/http/sync/AnnounceExample.java
@@ -1,6 +1,6 @@
 package com.bybit.api.examples.http.sync;
 
-import com.bybit.api.client.BybitAnnouncementRestClient;
+import com.bybit.api.client.BybitMarketDataRestClient;
 import com.bybit.api.client.domain.announcement.LanguageSymbol;
 import com.bybit.api.client.domain.announcement.request.AnnouncementInfoRequest;
 import com.bybit.api.client.service.BybitApiClientFactory;
@@ -8,7 +8,7 @@ import com.bybit.api.client.service.BybitApiClientFactory;
 public class AnnounceExample {
     public static void main(String[] args) {
         BybitApiClientFactory factory = BybitApiClientFactory.newInstance();
-        BybitAnnouncementRestClient client = factory.newAnnouncementRestClient();
+        var client = factory.newMarketDataRestClient();
 
         // Get Announcement
         var announcementInfoRequest = AnnouncementInfoRequest.builder().locale(LanguageSymbol.EN_US).build();
